long for pid.

+++ buffer-poll.c
@@ -41,8 +41,8 @@ buffer_poll(struct pollfd *pfd, struct buffer *in, struct buffer *out)
 {
 	ssize_t	n;
 
-	log_debug("buffer_poll (%d): fd=%d, revents=%d; out=%zu in=%zu",
-	    (int) getpid(),
+	log_debug("buffer_poll (%ld): fd=%d, revents=%d; out=%zu in=%zu",
+	    (long) getpid(),
 	    pfd->fd, pfd->revents, BUFFER_USED(out), BUFFER_USED(in));
 
 	if (pfd->revents & (POLLERR|POLLNVAL|POLLHUP))
